Seperated custom_columns resolvers from custom_columns model
